% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_table.R
\name{content_table}
\alias{content_table}
\title{Title}
\usage{
content_table(
  project_path = system.file("example_data/Zeeguu-Ecosystem/", package = "Rchitecture")
)
}
\arguments{
\item{project_path}{}
}
\value{
table
}
\description{
Title
}
