% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dependency_graph.R
\name{generate_dependency_graph}
\alias{generate_dependency_graph}
\title{NOT WORKING Visualize internal dependencies}
\usage{
generate_dependency_graph(project_path, path_prefix, dependencies_file)
}
\arguments{
\item{project_path}{path to the repository}

\item{path_prefix}{leading folders from absolute path to omit from visualization.}

\item{dependencies_file}{file with one dependency per line}
}
\description{
NOT WORKING Visualize internal dependencies
}
