% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_ast.R
\name{tokenize_ast}
\alias{tokenize_ast}
\title{Tokenizer for python ast}
\usage{
tokenize_ast(script_path, path_prefix = ".")
}
\arguments{
\item{script_path}{absolute paths from scripts to tokenize.}

\item{path_prefix}{leading folders from absolute path to omit from names.}
}
\value{
a named list, compatible with \code{quanteda::tokenize()}
}
\description{
Tokenizer for python ast
}
