% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_folder_type_and_size.R
\name{visualize_folder_type_and_size}
\alias{visualize_folder_type_and_size}
\title{Visualize folder type and size}
\usage{
visualize_folder_type_and_size(folder_path, path_prefix)
}
\arguments{
\item{folder_path}{path to the repository}

\item{path_prefix}{leading folders from absolute path to omit from visualization.}
}
\description{
Visualize folder type and size
}
