% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_table.R
\name{content_table}
\alias{content_table}
\title{Title}
\usage{
content_table(project_path, include_folders, path_prefix, .ignore_regex = NULL)
}
\arguments{
\item{project_path}{absolute path pointing to the files}

\item{include_folders}{should folders be included?}

\item{path_prefix}{leading part of the absolute file paths that is not meaningful for the analysis}

\item{.ignore_regex}{regex pattern for paths to exclude}
}
\value{
table
}
\description{
Title
}
