% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_dependency_graph.R
\name{internal_dependency_graph}
\alias{internal_dependency_graph}
\title{Visualize internal dependencies}
\usage{
internal_dependency_graph(project_path, path_prefix, .ignore_regex = NULL)
}
\arguments{
\item{project_path}{path to the repository}

\item{path_prefix}{leading folders from absolute path to omit from visualization.}

\item{.ignore_regex}{regex pattern for paths to exclude}
}
\description{
Visualize internal dependencies
}
